adding minimal settlement logic

+++ SplitExpenses/src/main/java/com/loosers/org/splitExpenses/model/GroupExpenseTable.java
@@ -2,10 +2,11 @@ package com.loosers.org.splitExpenses.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
 public class GroupExpenseTable {
     private String groupId;
-    private Map<String, Double> userExpenseMap;
+    private Map<String, BigDecimal> userExpenseMap;
 }
